pyproject.toml
[project]
name = "pyspark-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyspark-client==4.1.1",
    "pytest>=9.0.2",
]
